src/requirements.txt
discord.py==2.0.1
psutil==5.9.1